complete css flex landing page
It's mostly in a state I'm pleased with. I do want to return
to find a way to make the footer scale to the page without
forcing scrolling. I think my page can be cleaned up in general
but the building of the HTML is easily the most complex
html file I've ever made.
Thumbs up.